feat: edit ticket tier prices
